Ignore mismatched libraries if they are upgraded
Previously, if we tried to upgrade one of the watched libraries, the
mismatch detector might flag it as, itself, using an old version of
itself. This is because if an older version of this library were used
in another subtree, it might be shown as referencing an older version.
This just ignores looking at watched libraries when they themselves are
being upgraded.